+++ HISTORY.md
@@ -73,6 +73,9 @@
     * Greek, thanks, HM100!
     * Armenian, thanks to the contributors!
 
+### 2023-07-11
+* blocks, symbols, gui: design overhaul of variadic input slots
+
 ### 2023-07-10
 * new "Tiles" library, divides the stage into sub-regions in each of which to perform an action 
 * new "Arcs" library, turns sprites by a delta of degrees moving them at a given radius